add test for double/float64 field

+++ flow/connectors/clickhouse/qvalue_convert.go
@@ -9,7 +9,7 @@ import (
 var clickhouseTypeToQValueKindMap = map[string]qvalue.QValueKind{
 	"INT":           qvalue.QValueKindInt32,
 	"Int64":         qvalue.QValueKindInt64,
-	"FLOAT":         qvalue.QValueKindFloat64,
+	"Float64":       qvalue.QValueKindFloat64,
 	"DOUBLE":        qvalue.QValueKindFloat64,
 	"REAL":          qvalue.QValueKindFloat64,
 	"VARCHAR":       qvalue.QValueKindString,
